+++ World-Builders-Handbook-2e/src/main/java/vandyke/Utilites/Formulas.java
@@ -2,10 +2,10 @@ package vandyke.Utilites;
 
 public class Formulas {
 
-    public static Float Interpolate(Float percentage, Float lowerValue, Float upperValue) {
-        Float difference = upperValue - lowerValue;
+    public static Double Interpolate(Double percentage, Double lowerValue, Double upperValue) {
+        Double difference = upperValue - lowerValue;
         //System.out.println("difference: " + difference);
-        Float interpolatedResult = difference * percentage;
+        Double interpolatedResult = difference * percentage;
         //System.out.println("interpolated result: " + interpolatedResult);
         return interpolatedResult;
     }
